+++ scripts/channelbalance.py
@@ -76,6 +76,8 @@ if __name__ == '__main__':
     colorBarOutline=ImageColor.getrgb("#770044")
     colorBarFilled=ImageColor.getrgb("#aa3377")
     colorBarEmpty=ImageColor.getrgb("#202020")
+    width=480
+    height=320
     sleepInterval=1800
     pageSize=8
     # Override config
@@ -96,6 +98,10 @@ if __name__ == '__main__':
             colorBarFilled = ImageColor.getrgb(config["colorBarFilled"])
         if "colorBarEmpty" in config:
             colorBarEmpty = ImageColor.getrgb(config["colorBarEmpty"])
+        if "width" in config:
+            width = int(config["width"])
+        if "height" in config:
+            height = int(config["height"])
         if "sleepInterval" in config:
             sleepInterval = int(config["sleepInterval"])
             sleepInterval = 300 if sleepInterval < 300 else sleepInterval # minimum 5 minutes, access others
@@ -121,5 +127,5 @@ if __name__ == '__main__':
             lastidx = (pagenum*pageSize)-1
             if lastidx > channelcount-1:
                 lastidx = channelcount-1
-            createimage(channels, firstidx, lastidx, pagenum, pageSize)
+            createimage(channels, firstidx, lastidx, pagenum, pageSize, width, height)
         time.sleep(sleepInterval)