+++ CHANGELOG.md
@@ -11,6 +11,7 @@ New Panels and Enhancements
 
 General
 - Updated macaroon creation to allow querying ForwardingHistory
+- Dimensions for any panel can be set via "width" and "height" in config file
 
 ## 1.0.0